+++ lib/src/http_client.dart
@@ -29,26 +29,37 @@ class HTTPClient {
     }
   }
 
-  int _connectTimeOut = 500;
-  int _receiveTimeOut = 1000;
-  int _sendTimeOut = 500;
+  int _connectTimeOut = 5000;
+  int _receiveTimeOut = 10000;
+  int _sendTimeOut = 5000;
 
   Future<Response> get(String url,
       {Map<String, String> headers = const {},
       Iterable<Cookie> cookies = const []}) async {
-    Response resp = await dio.get(url,
-        options: Options(
-            connectTimeout: _connectTimeOut,
-            receiveTimeout: _receiveTimeOut,
-            sendTimeout: _sendTimeOut,
-            headers: headers,
-            cookies: cookies,
-            followRedirects: false,
-            validateStatus: (status) {
-              // 500 would be caught later,
-              // this is for the sake of exception consistency.
-              return status <= 500;
-            }));
+    Response resp;
+    try {
+      resp = await dio.get(url,
+          options: Options(
+              connectTimeout: _connectTimeOut,
+              receiveTimeout: _receiveTimeOut,
+              sendTimeout: _sendTimeOut,
+              headers: headers,
+              cookies: cookies,
+              followRedirects: false,
+              validateStatus: (status) {
+                // 500 would be caught later,
+                // this is for the sake of exception consistency.
+                return status <= 500;
+              }));
+    } catch (e) {
+      if (e.toString().startsWith(
+          "DioError [DioErrorType.DEFAULT]: NoSuchMethodError: The method 'startsWith' was called on null.")) {
+        throw WrongCredentialsException();
+      } else {
+        rethrow;
+      }
+    }
+
     _catchHttpExceptions(resp, 'HTTPClient.get');
     return resp;
   }